Move Prometheus config into internal/.
To reduce a number of identifiers exported by
prometheus package.
